+++ app/problem_register/forms.py
@@ -1,0 +1,16 @@
+from django import forms
+from leaflet.forms.widgets import LeafletWidget
+
+from problem_register.models import ProblemLabel
+
+
+class ProblemLabelForm(forms.ModelForm):
+    class Meta:
+        model = ProblemLabel
+        fields = ('description', 'geom')
+        labels = {
+            'description': '',
+            'geom': '',
+        }
+        widgets = {'description': forms.Textarea(attrs={"rows": 5, "cols": 10, 'placeholder': 'Описание проблемы'}),
+                   'geom': LeafletWidget()}
